+++ src/main/kotlin/dev/mayaqq/cynosure/events/entity/player/PlayerEvents.kt
@@ -19,3 +19,5 @@ public sealed class PlayerConnectionEvent(public val player: ServerPlayer) : Eve
 }
 
 public class PlayerRestoreEvent(public val player: Player, public val alive: Boolean) : Event
+
+public class PlayerRespawnEvent(public val newPlayer: ServerPlayer, public val oldPlayer: ServerPlayer, public val alive: Boolean) : Event