+++ CHANGELOG.md
@@ -1,6 +1,5 @@
 # Cynosure `@VERSION@` `1.20.1` `Fabric & Forge`
 
 ### Changelog:
-- ServerChatEvent for messages sent to the player on the Server!
-- Made damage event cancel any incoming damage if return value is 0.0F
-- Fixed Forge Loot Table Modifiers being broken
+- Player Respawn Event
+- Fixed Access Wideners on Fabric